requirements.txt
pytesseract
pyrogram
Pillow